fix wrong naming
