+++ martin/src/mbtiles/source.rs
@@ -1,5 +1,168 @@
-mod error;
-pub use error::MbtilesError;
+use std::fmt::{Debug, Formatter};
+use std::io;
+use std::path::PathBuf;
+use std::sync::Arc;
 
-mod source;
-pub use source::MbtilesSource;
+use async_trait::async_trait;
+use log::trace;
+use martin_tile_utils::{TileCoord, TileInfo};
+use mbtiles::MbtilesPool;
+use tilejson::TileJSON;
+
+use crate::config::file::ConfigFileError::{InvalidMetadata, IoError};
+use crate::config::file::ConfigFileResult;
+use crate::source::{TileData, TileInfoSource, UrlQuery};
+use crate::{MartinResult, Source};
+
+#[derive(Clone)]
+pub struct MbtSource {
+    id: String,
+    mbtiles: Arc<MbtilesPool>,
+    tilejson: TileJSON,
+    tile_info: TileInfo,
+}
+
+impl Debug for MbtSource {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "MbtSource {{ id: {}, path: {:?} }}",
+            self.id,
+            self.mbtiles.as_ref()
+        )
+    }
+}
+
+impl MbtSource {
+    pub async fn new(id: String, path: PathBuf) -> ConfigFileResult<Self> {
+        let mbt = MbtilesPool::open_readonly(&path)
+            .await
+            .map_err(|e| io::Error::other(format!("{e:?}: Cannot open file {}", path.display())))
+            .map_err(|e| IoError(e, path.clone()))?;
+
+        let meta = mbt
+            .get_metadata()
+            .await
+            .map_err(|e| InvalidMetadata(e.to_string(), path))?;
+
+        Ok(Self {
+            id,
+            mbtiles: Arc::new(mbt),
+            tilejson: meta.tilejson,
+            tile_info: meta.tile_info,
+        })
+    }
+}
+
+#[async_trait]
+impl Source for MbtSource {
+    fn get_id(&self) -> &str {
+        &self.id
+    }
+
+    fn get_tilejson(&self) -> &TileJSON {
+        &self.tilejson
+    }
+
+    fn get_tile_info(&self) -> TileInfo {
+        self.tile_info
+    }
+
+    fn clone_source(&self) -> TileInfoSource {
+        Box::new(self.clone())
+    }
+
+    fn benefits_from_concurrent_scraping(&self) -> bool {
+        // If we copy from one local file to another, we are likely not bottlenecked by CPU
+        false
+    }
+
+    async fn get_tile(
+        &self,
+        xyz: TileCoord,
+        _url_query: Option<&UrlQuery>,
+    ) -> MartinResult<TileData> {
+        if let Some(tile) = self
+            .mbtiles
+            .get_tile(xyz.z, xyz.x, xyz.y)
+            .await
+            .map_err(|_| MbtilesError::AcquireConnError(self.id.clone()))?
+        {
+            Ok(tile)
+        } else {
+            trace!(
+                "Couldn't find tile data in {}/{}/{} of {}",
+                xyz.z, xyz.x, xyz.y, &self.id
+            );
+            Ok(Vec::new())
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::BTreeMap;
+    use std::path::PathBuf;
+
+    use indoc::indoc;
+
+    use crate::config::file::mbtiles::MbtConfig;
+    use crate::config::file::{FileConfigEnum, FileConfigSource, FileConfigSrc};
+
+    #[test]
+    fn parse() {
+        let cfg = serde_yaml::from_str::<FileConfigEnum<MbtConfig>>(indoc! {"
+            paths:
+              - /dir-path
+              - /path/to/file2.ext
+              - http://example.org/file.ext
+            sources:
+                pm-src1: /tmp/file.ext
+                pm-src2:
+                  path: /tmp/file.ext
+                pm-src3: https://example.org/file3.ext
+                pm-src4:
+                  path: https://example.org/file4.ext
+        "})
+        .unwrap();
+        let res = cfg.finalize("");
+        assert!(res.is_empty(), "unrecognized config: {res:?}");
+        let FileConfigEnum::Config(cfg) = cfg else {
+            panic!();
+        };
+        let paths = cfg.paths.clone().into_iter().collect::<Vec<_>>();
+        assert_eq!(
+            paths,
+            vec![
+                PathBuf::from("/dir-path"),
+                PathBuf::from("/path/to/file2.ext"),
+                PathBuf::from("http://example.org/file.ext"),
+            ]
+        );
+        assert_eq!(
+            cfg.sources,
+            Some(BTreeMap::from_iter(vec![
+                (
+                    "pm-src1".to_string(),
+                    FileConfigSrc::Path(PathBuf::from("/tmp/file.ext"))
+                ),
+                (
+                    "pm-src2".to_string(),
+                    FileConfigSrc::Obj(FileConfigSource {
+                        path: PathBuf::from("/tmp/file.ext"),
+                    })
+                ),
+                (
+                    "pm-src3".to_string(),
+                    FileConfigSrc::Path(PathBuf::from("https://example.org/file3.ext"))
+                ),
+                (
+                    "pm-src4".to_string(),
+                    FileConfigSrc::Obj(FileConfigSource {
+                        path: PathBuf::from("https://example.org/file4.ext"),
+                    })
+                ),
+            ]))
+        );
+    }
+}
